+++ src/constant.ts
@@ -1,5 +1,6 @@
-export const DIALOG_FLOW_ACTION = '__action__';
-export const DIALOG_FLOW_INTENT = '__intent__';
+export const DIALOG_FLOW_ACTION = '__dfAction__';
+export const DIALOG_FLOW_INTENT = '__dfIntent__';
+export const DIALOG_FLOW_PARAMS = '__dfParams__';
 export const nestMetadata = { COMPONENTS: 'components' };
 export const PATH_METADATA = 'path';
 export const METHOD_METADATA = 'method';